Reimplement goal indicator lines
