+++ Assets/Scripts/Goal.cs
@@ -10,7 +10,6 @@ public class Goal : MonoBehaviour {
     private static readonly float SNAP_THRESHOLD = 1.5f;
     private static readonly float X_MIN = -2f, X_MAX = 2f, Y_MIN = 0.5f, Y_MAX = 3.5f, Z_MIN = -8f, Z_MAX =2f;
     private static readonly float PLAYER_1_ROTATION = 180f, PLAYER_2_ROTATION = 0;
-    private static readonly int LENGTH_LINERENDERER = 2;
 
     private Vector3 parentPos, newPos, lastSafePos;
     private int playerScore;
@@ -24,7 +23,6 @@ public class Goal : MonoBehaviour {
     }
 
     private Utils.PlayerNumber playerNumber;
-    private LineRenderer goalIndicator;
 
     // Start is called before the first frame update
     void Start() {
@@ -34,7 +32,6 @@ public class Goal : MonoBehaviour {
             playerNumber = Utils.PlayerNumber.TWO;
         }
         ResetGoal();
-        InitGoalIndicator();
         GameManager.Instance.RestartEvent.AddListener(ResetGoal);
     }
 
@@ -48,22 +45,6 @@ public class Goal : MonoBehaviour {
         }
     }
 
-    private void InitGoalIndicator() {
-        goalIndicator = gameObject.AddComponent<LineRenderer>();
-        goalIndicator.material = new Material(Shader.Find("Sprites/Default"));
-        goalIndicator.widthMultiplier = 0.025f;
-        goalIndicator.positionCount = LENGTH_LINERENDERER;
-
-        Color c1 = Color.white;
-        Color c2 = Color.green;
-        float alpha = 0.5f;
-        Gradient gradient = new Gradient();
-        gradient.SetKeys(
-            new GradientColorKey[] { new GradientColorKey(c1, 0.0f), new GradientColorKey(c2, 0.5f), new GradientColorKey(c1, 1.0f) },
-            new GradientAlphaKey[] { new GradientAlphaKey(alpha, 0.0f), new GradientAlphaKey(alpha, 1.0f) });
-        goalIndicator.colorGradient = gradient;
-    }
-
     public void ResetGoal() {
         InitRotationAndOffset();
         goalState = GoalState.FOLLOWING;
@@ -72,7 +53,6 @@ public class Goal : MonoBehaviour {
     // Update is called once per frame
     void Update() {
         HandleGoalPosition();
-        HandleGoalIndicator();
     }
     
     private void HandleGoalPosition() {
@@ -138,16 +118,5 @@ public class Goal : MonoBehaviour {
 
     private void SwitchGoalState() {
         goalState = (goalState == GoalState.FOLLOWING) ? GoalState.STATIONARY : GoalState.TRANSITION;
-    }
-
-    private void HandleGoalIndicator() {
-        Vector3 lrOffsetL = new Vector3(-0.5f, 0.45f, 0);
-        Vector3 lrOffsetR = new Vector3(0.5f, 0.45f, 0);
-
-        goalIndicator.SetVertexCount(4);
-        goalIndicator.SetPosition(0, -transform.forward * LENGTH_LINERENDERER + (transform.position + lrOffsetL));
-        goalIndicator.SetPosition(1, transform.position + lrOffsetL);
-        goalIndicator.SetPosition(2, transform.position + lrOffsetR);
-        goalIndicator.SetPosition(3, -transform.forward * LENGTH_LINERENDERER + (transform.position + lrOffsetR));
     }
 }
